requirements.txt
evernote3==1.25.14